pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jupyter_sec_firewall"
version = "0.1.0"
description = "A real-time Jupyter execution firewall that intercepts and blocks malicious commands."
authors = [{name = "Akil", email = "akhilanand32@gmail.com"}]
dependencies = [
    "jupyter_server>=2.0.0",
    "tornado",
    "websocket-client>=1.0",  # required by test_ws.py integration test
]
readme = "README.md"
requires-python = ">=3.8"

[project.entry-points."jupyter_server.extension"]
jupyter_sec_firewall = "jupyter_sec_firewall.extension"

[tool.hatch.build.targets.wheel]
packages = ["jupyter_sec_firewall"]

[tool.hatch.build.targets.wheel.shared-data]
"jupyter_sec_firewall/jupyter_config/jupyter_server_config.d/jupyter_sec_firewall.json" = "etc/jupyter/jupyter_server_config.d/jupyter_sec_firewall.json"
